requirements.txt
streamlit-1.12.0
plyer-2.0.0
pytube-12.1.0.